fix(git): parse sparse and cquoted paths on the SSH relay
Share parseWorktreeList and parseUnmergedEntry in src/shared so SSH
cannot drift from local: porcelain sparse sets optional isSparse, and
unmerged paths go through decodeGitCQuotedPath with the desktop
ENOENT/ENOTDIR probe.

Fixes #18280
