+++ src/relay/git-handler-utils.ts
@@ -1,16 +1,16 @@
 /**
- * Pure parsing helpers extracted from git-handler.ts.
+ * Git porcelain/diff helpers extracted from git-handler.ts.
  *
  * Why: oxlint max-lines requires files to stay under 300 lines.
- * These functions have no side-effects and depend only on their arguments,
- * making them easy to test independently.
+ * Branch-diff helpers are pure; `parseUnmergedEntry` is a shared re-export
+ * that probes the worktree filesystem.
  */
-import { existsSync } from 'node:fs'
 import * as path from 'node:path'
-import { decodeGitCQuotedPath } from '../shared/git-cquoted-path'
 import { isBinaryBuffer } from '../shared/binary-buffer'
 import type { GitLineStats } from '../shared/git-uncommitted-line-stats'
 export { isUnsupportedWorktreeListZError } from '../shared/git-worktree-command-capabilities'
+export { parseWorktreeList } from '../shared/git-worktree-list-parser'
+export { parseUnmergedEntry } from '../shared/git-unmerged-entry'
 
 export function parseBranchStatusChar(char: string): string {
   switch (char) {
@@ -26,76 +26,6 @@ export function parseBranchStatusChar(char: string): string {
       return 'copied'
     default:
       return 'modified'
-  }
-}
-
-export function parseConflictKind(xy: string): string | null {
-  switch (xy) {
-    case 'UU':
-      return 'both_modified'
-    case 'AA':
-      return 'both_added'
-    case 'DD':
-      return 'both_deleted'
-    case 'AU':
-      return 'added_by_us'
-    case 'UA':
-      return 'added_by_them'
-    case 'DU':
-      return 'deleted_by_us'
-    case 'UD':
-      return 'deleted_by_them'
-    default:
-      return null
-  }
-}
-
-/**
- * Parse a single unmerged entry line from porcelain v2 output.
- * Returns null if the entry should be skipped (e.g. submodule conflicts).
- */
-export function parseUnmergedEntry(
-  worktreePath: string,
-  line: string
-): Record<string, unknown> | null {
-  const parts = line.split(' ')
-  const xy = parts[1]
-  const modeStage1 = parts[3]
-  const modeStage2 = parts[4]
-  const modeStage3 = parts[5]
-  const filePath = parts.slice(10).join(' ')
-  if (!filePath) {
-    return null
-  }
-
-  // Skip submodule conflicts (mode 160000)
-  if ([modeStage1, modeStage2, modeStage3].some((m) => m === '160000')) {
-    return null
-  }
-
-  const conflictKind = parseConflictKind(xy)
-  if (!conflictKind) {
-    return null
-  }
-
-  let status: string = 'modified'
-  if (conflictKind === 'both_deleted') {
-    status = 'deleted'
-  } else if (conflictKind !== 'both_modified' && conflictKind !== 'both_added') {
-    try {
-      status = existsSync(path.join(worktreePath, filePath)) ? 'modified' : 'deleted'
-    } catch {
-      // Why: defaulting to 'modified' on fs error is the least misleading option
-      status = 'modified'
-    }
-  }
-
-  return {
-    path: filePath,
-    area: 'unstaged',
-    status,
-    conflictKind,
-    conflictStatus: 'unresolved'
   }
 }
 
@@ -131,100 +61,6 @@ export function parseBranchDiff(
     }
   }
   return entries
-}
-
-// ─── Worktree parsing ────────────────────────────────────────────────
-
-export function parseWorktreeList(
-  output: string,
-  options: { nulDelimited?: boolean } = {}
-): Record<string, unknown>[] {
-  const worktrees: Record<string, unknown>[] = []
-  const blocks = options.nulDelimited ? splitNulWorktreeList(output) : splitLineWorktreeList(output)
-
-  for (const lines of blocks) {
-    if (lines.length === 0) {
-      continue
-    }
-    let wtPath = ''
-    let head = ''
-    let branch = ''
-    let isBare = false
-    let locked = false
-    let lockReason = ''
-    let prunable = false
-    let prunableReason = ''
-
-    for (const line of lines) {
-      if (line.startsWith('worktree ')) {
-        wtPath = line.slice('worktree '.length)
-      } else if (line.startsWith('HEAD ')) {
-        head = line.slice('HEAD '.length)
-      } else if (line.startsWith('branch ')) {
-        branch = line.slice('branch '.length)
-      } else if (line === 'bare') {
-        isBare = true
-      } else if (line === 'locked' || line.startsWith('locked ')) {
-        locked = true
-        const rawReason = line.slice('locked'.length).trim()
-        lockReason = options.nulDelimited ? rawReason : decodeGitCQuotedPath(rawReason)
-      } else if (line === 'prunable' || line.startsWith('prunable ')) {
-        // Why: Git ≥ 2.36 flags registrations whose directory is gone; ignoring
-        // it surfaces the stale worktree as a live workspace (issue #8389).
-        prunable = true
-        const rawReason = line.slice('prunable'.length).trim()
-        prunableReason = options.nulDelimited ? rawReason : decodeGitCQuotedPath(rawReason)
-      }
-    }
-
-    if (wtPath) {
-      worktrees.push({
-        path: wtPath,
-        head,
-        branch,
-        isBare,
-        ...(locked ? { locked: true } : {}),
-        ...(lockReason ? { lockReason } : {}),
-        ...(prunable ? { prunable: true } : {}),
-        ...(prunableReason ? { prunableReason } : {}),
-        isMainWorktree: worktrees.length === 0
-      })
-    }
-  }
-  return worktrees
-}
-
-function splitLineWorktreeList(output: string): string[][] {
-  return output
-    .trim()
-    .split(/\r?\n\r?\n/)
-    .map((block) => block.trim().split(/\r?\n/))
-}
-
-function splitNulWorktreeList(output: string): string[][] {
-  if (!output.includes('\0')) {
-    return splitLineWorktreeList(output)
-  }
-
-  const blocks: string[][] = []
-  let currentBlock: string[] = []
-
-  for (const field of output.split('\0')) {
-    if (field) {
-      currentBlock.push(field)
-      continue
-    }
-    if (currentBlock.length > 0) {
-      blocks.push(currentBlock)
-      currentBlock = []
-    }
-  }
-
-  if (currentBlock.length > 0) {
-    blocks.push(currentBlock)
-  }
-
-  return blocks
 }
 
 // ─── Binary / blob helpers ───────────────────────────────────────────
